feat: fixed typescript interface
