+++ ambient.d.ts
@@ -1,2 +1,2 @@
-declare module 'whatsapp-web.js';
+// declare module 'whatsapp-web.js';
 declare module 'qrcode-terminal';
