+++ src/app/interfaces/Cep.ts
@@ -1,0 +1,12 @@
+interface IResponse {
+  data: IServerData;
+}
+
+interface IServerData {
+  cep: string;
+  logradouro: string;
+  bairro: string;
+  uf: string;
+}
+
+export { IResponse, IServerData };
